api/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Google Gemini SDK
google-generativeai>=0.8.0

# Vector database
qdrant-client>=1.7.0

# PostgreSQL
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.25
asyncpg>=0.29.0

# Configuration
python-dotenv>=1.0.0
pydantic-settings>=2.1.0

# HTTP client
httpx>=0.26.0

# Markdown processing
markdown>=3.5.0
beautifulsoup4>=4.12.0

# Utilities
tiktoken>=0.5.0
python-frontmatter>=1.1.0
